# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Employee API
  version: 1.0.0
paths:
  /employees/{id}/skills:
    put:
      summary: Update skills for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skills:
                  type: array
                  items:
                    type: object
                    properties:
                      experience:
                        type: string
                      skillName:
                        type: string
                      level:
                        type: string
                    required:
                      - experience
                      - skillName
                      - level
              required:
                - skills
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

  /employees/{id}/weekSchedules:
    put:
      summary: Update week schedules for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                weekSchedules:
                  type: array
                  items:
                    type: object
                    properties:
                      day:
                        type: string
                      startSchedule:
                        type: string
                      endSchedule:
                        type: string
                    required:
                      - day
                      - startSchedule
                      - endSchedule
              required:
                - weekSchedules
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

  /employees/{id}/foreignLanguages:
    put:
      summary: Update foreign languages for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                foreignLanguages:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      level:
                        type: string
                    required:
                      - name
                      - level
              required:
                - foreignLanguages
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        skills:
          type: array
          items:
            type: object
            properties:
              experience:
                type: string
              skillName:
                type: string
              level:
                type: string
        weekSchedules:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              startSchedule:
                type: string
              endSchedule:
                type: string
        foreignLanguages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              level:
                type: string